+++ test/unit/v2/connectors/test_confluence.py
@@ -11,10 +11,10 @@ def test_connection_config_multiple_auth():
     with pytest.raises(ValidationError):
         ConfluenceConnectionConfig(
             access_config=ConfluenceAccessConfig(
-                api_token="api_token",
-                access_token="access_token",
+                password="api_token",
+                token="access_token",
             ),
-            user_email="user_email",
+            username="user_email",
             url="url",
         )
 
@@ -26,14 +26,14 @@ def test_connection_config_no_auth():
 
 def test_connection_config_basic_auth():
     ConfluenceConnectionConfig(
-        access_config=ConfluenceAccessConfig(api_token="api_token"),
+        access_config=ConfluenceAccessConfig(password="api_token"),
         url="url",
-        user_email="user_email",
+        username="user_email",
     )
 
 
 def test_connection_config_pat_auth():
     ConfluenceConnectionConfig(
-        access_config=ConfluenceAccessConfig(access_token="access_token"),
+        access_config=ConfluenceAccessConfig(token="access_token"),
         url="url",
     )